Move ConstructionStrategy and ConstructionData types
