+++ emergence_lib/src/structures/structure_manifest.rs
@@ -2,14 +2,14 @@
 
 use crate::{
     asset_management::manifest::{loader::IsRawManifest, Id, Manifest},
-    crafting::components::{ActiveRecipe, InputInventory, RawActiveRecipe},
-    items::{item_manifest::Item, slot::ItemSlot},
+    construction::{ConstructionData, ConstructionStrategy, RawConstructionStrategy},
+    crafting::components::{ActiveRecipe, RawActiveRecipe},
+    items::item_manifest::Item,
     organisms::{OrganismId, OrganismVariety, RawOrganismVariety},
-    terrain::terrain_manifest::Terrain,
 };
 use bevy::{
     reflect::{FromReflect, Reflect, TypeUuid},
-    utils::{Duration, HashMap, HashSet},
+    utils::HashMap,
 };
 use serde::{Deserialize, Serialize};
 
@@ -74,78 +74,6 @@ impl From<RawStructureData> for StructureData {
             construction_strategy: raw.construction_strategy.into(),
             max_workers: raw.max_workers,
             footprint: raw.footprint.unwrap_or_default(),
-        }
-    }
-}
-
-/// How new structures of this sort can be built.
-///
-/// For structures that are part of a `Lifecycle`, this should generally be the same for all of them.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub enum ConstructionStrategy {
-    /// Follows the construction strategy of another structure.
-    Seedling(Id<Structure>),
-    /// This structure can be built directly.
-    Direct(ConstructionData),
-}
-
-/// The data contained in a [`ConstructionStrategy::Direct`] variant.
-#[derive(Debug, Default, Clone, PartialEq, Serialize, Deserialize)]
-pub struct ConstructionData {
-    /// The amount of work by units required to complete the construction of this building.
-    ///
-    /// If this is [`None`], no work will be needed at all.
-    pub work: Option<Duration>,
-    /// The set of items needed to create a new copy of this structure
-    pub materials: InputInventory,
-    /// The set of terrain types that this structure can be built on
-    pub allowed_terrain_types: HashSet<Id<Terrain>>,
-}
-
-/// The unprocessed equivalent of [`ConstructionStrategy`].
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub enum RawConstructionStrategy {
-    /// Follows the construction strategy of another structure.
-    Seedling(String),
-    /// This structure can be built directly.
-    Direct {
-        /// The amount of work (in seconds) by units required to complete the construction of this building.
-        ///
-        /// If this is [`None`], no work will be needed at all.
-        work: Option<f32>,
-        /// The set of items needed to create a new copy of this structure
-        materials: HashMap<String, u32>,
-        /// The set of terrain types that this structure can be built on
-        allowed_terrain_types: HashSet<String>,
-    },
-}
-
-impl From<RawConstructionStrategy> for ConstructionStrategy {
-    fn from(raw: RawConstructionStrategy) -> Self {
-        match raw {
-            RawConstructionStrategy::Seedling(seedling_name) => {
-                ConstructionStrategy::Seedling(Id::from_name(seedling_name))
-            }
-            RawConstructionStrategy::Direct {
-                work,
-                materials,
-                allowed_terrain_types,
-            } => {
-                let inventory = materials
-                    .into_iter()
-                    .map(|(item_name, count)| ItemSlot::new(Id::from_name(item_name), count))
-                    .collect();
-
-                let materials = InputInventory::Exact { inventory };
-                ConstructionStrategy::Direct(ConstructionData {
-                    work: work.map(Duration::from_secs_f32),
-                    materials,
-                    allowed_terrain_types: allowed_terrain_types
-                        .into_iter()
-                        .map(Id::from_name)
-                        .collect(),
-                })
-            }
         }
     }
 }
